Move code to cortex-m7
