+++ libcpu/arm/cortex-m7/context_gcc.S
@@ -17,6 +17,8 @@
  */
 /*@{*/
 
+#include <rtconfig.h>
+
 .cpu cortex-m4
 .syntax unified
 .thumb
@@ -150,6 +152,14 @@ switch_to_thread:
     CMP     r3,  #0             /* if(flag_r3 != 0) */
     IT      NE
     BICNE   lr, lr, #0x10       /* lr &= ~(1 << 4), set FPCA. */
+#endif
+
+#if defined (RT_USING_MEM_PROTECTION)
+    PUSH    {r0-r3, r12, lr}
+    LDR     r1, =rt_current_thread
+    LDR     r0, [r1]
+    BL      rt_hw_mp_table_switch
+    POP     {r0-r3, r12, lr}
 #endif
 
 pendsv_exit:
